requirements.txt
pandas
numpy
scikit-learn
nltk
spacy
streamlit
joblib
pdfplumber
python-docx
